+++ src/flipper/component.js
@@ -500,6 +500,10 @@ Component.prototype = {
             result = window[element.getAttribute('model-key')];
         }
 
+        if (!modelId && element.hasAttribute('model-id')) {
+            Flipper.unlinkSpace(element.getAttribute('model-id'));
+        }
+
         return Promise.resolve(result).then(function(model) {
             if (model !== undefined) {
                 element.model = model;
@@ -607,6 +611,8 @@ Component.prototype = {
         if (!Flipper.useNative) {
             Flipper.parse(element);
         }
+
+        triggerExternalLifeEvent(element, 'rendered');
 
         var result = tryCallLifeCycleEvent(element, 'ready');
         return Promise.resolve(result).then(function() {